Add empty input check, keep showing prompt if empty imput

+++ e-learning-env-init.sh
@@ -28,11 +28,11 @@ done
 
 echo "You selected $bashBranch"
 
-echo "Clone the base environment docker repo: git clone -b $bashBranch --single-branch https://github.com/catalyst/docker_moodle.git ."
-git clone -b $bashBranch --single-branch https://github.com/catalyst/docker_moodle.git .
+echo "Clone the base environment docker repo: git clone -b $bashBranch https://github.com/catalyst/docker_moodle.git ."
+git clone -b $bashBranch https://github.com/catalyst/docker_moodle.git . && \
 
-echo "docker-composer down and create the containers again:\n docker-compose -f 'docker-compose.yml' down && docker-compose up -d --force-recreate --build"
-docker-compose -f 'docker-compose.yml' down && docker-compose up -d --force-recreate --build
+echo "docker-composer down and create the containers again: docker-compose -f 'docker-compose.yml' down && docker-compose up -d --force-recreate --build"
+docker-compose -f 'docker-compose.yml' down && docker-compose up -d --force-recreate --build  && \
 
 echo "update owner and user group for ./siteroot, run: chown ${USER}:${USER} ./siteroot"
 sudo chown ${USER}:${USER} ./siteroot
@@ -40,22 +40,31 @@ cd siteroot
 
 promptInitRepoURL='Please enter your site repo git URL:'
 echo $promptInitRepoURL
-read repoURL
-echo "$repoURL is entered.\n"
+while read repoURL; do
+    if [[ ! -z ${repoURL} ]]; then
+        echo "$repoURL is entered."
+        break
+    fi
+done
 
 promptInitRepoBranch='Please enter your site repo branch name:'
 echo $promptInitRepoBranch
-read repoBranch
-echo "$repoBranch is entered.\n"
+while read repoBranch; do
+    if [[ ! -z ${repoBranch} ]]; then
+        echo "$repoBranch is entered."
+        break
+    fi
+done
 
-echo "Clone moodle repo: git clone --single-branch -b ${repoBranch} ${repoURL} ."
-git clone -b ${repoBranch} --single-branch ${repoURL} .
+
+echo "Clone moodle repo: git clone -b ${repoBranch} ${repoURL} ."
+git clone -b ${repoBranch} ${repoURL} .  && \
 
 echo "Copy config file: cp moodle-config siteroot/config.php"
-cp ../moodle-config config.php
+cp ../moodle-config config.php  && \
 
 echo "Update git submodule: git submodule update --init --recursive --jobs 16"
-git submodule update --init --recursive --jobs 16
+git submodule update --init --recursive --jobs 16  && \
 
 cd ../
 
@@ -65,7 +74,27 @@ if [[ ! -f "${shellPWD}/docker-compose.yml" ]]; then
     exit
 fi
 
-webcont="$(docker-compose ps -q moodle)"
-wget -O - https://raw.githubusercontent.com/jasonliancatalyst/docker-env-init/master/container-setup.sh | docker exec -i $webcont bash -
+promptInitInsideContainer='Do you wish to run init jobs (composer install/phpunit init) inside container? [Y/N]'
+echo $promptInitInsideContainer
+while read answerInitInsideContainer; do
+    if [[ ! -z $answerInitInsideContainer ]]; then
+        echo "$answerInitInsideContainer is entered."
+        break
+    fi
+done
 
-sudo chown ${USER}:${USER} ./siteroot
+if [[ ${answerInitInsideContainer,,} == "y" ]]
+then
+    echo "Proceeding...."
+
+    webcont="$(docker-compose ps -q moodle)" && \
+    wget -O - https://raw.githubusercontent.com/jasonliancatalyst/docker-env-init/master/container-setup.sh | docker exec -i $webcont bash -
+
+    sudo chown ${USER}:${USER} ./siteroot
+else
+    echo "You can run it manually:"
+    echo "webcont=\"$(docker-compose ps -q moodle)\" && \\"
+    echo "wget -O - https://raw.githubusercontent.com/jasonliancatalyst/docker-env-init/master/container-setup.sh | docker exec -i $webcont bash -"
+    echo "Exiting...."
+    exit
+fi
